Error reporting

+++ DiskSmartInfo.functions.ps1
@@ -2,7 +2,7 @@ function Get-DiskSmartInfo
 {
     [CmdletBinding(DefaultParameterSetName='ComputerName')]
     Param(
-        [Parameter(ParameterSetName='ComputerName')]
+        [Parameter(Position=0,ParameterSetName='ComputerName')]
         [string[]]$ComputerName,
         [Parameter(ParameterSetName='CimSession')]
         [CimSession[]]$CimSession,
@@ -11,43 +11,83 @@ function Get-DiskSmartInfo
         [switch]$SilenceIfNotInWarningOrCriticalState
     )
 
+    $Script:ErrorCreatingCimSession = @()
+    $Script:ErrorAccessingCimSession = @()
+    $Script:ErrorAccessingClass = @()
+
+    # ComputerName
     if ($ComputerName)
     {
         try
         {
-            $cimSessions = New-CimSession -ComputerName $ComputerName
+            if ($DebugPreference -eq 'Continue')
+            {
+                $cimSessions = New-CimSession -ComputerName $ComputerName
+            }
+            else
+            {
+                $cimSessions = New-CimSession -ComputerName $ComputerName -ErrorVariable Script:ErrorCreatingCimSession -ErrorAction SilentlyContinue
+            }
 
             foreach ($cim in $cimSessions)
             {
                 inGetDiskSmartInfo `
-                    -Session $cim `
-                    -ShowConvertedData:$ShowConvertedData `
-                    -CriticalAttributesOnly:$CriticalAttributesOnly `
-                    -SilenceIfNotInWarningOrCriticalState:$SilenceIfNotInWarningOrCriticalState
-            }
-        }
-        finally
-        {
-            Remove-CimSession -CimSession $cimSessions
-        }
-    }
-    elseif ($CimSession)
-    {
-        foreach ($cim in $CimSession)
-        {
-            inGetDiskSmartInfo `
                 -Session $cim `
                 -ShowConvertedData:$ShowConvertedData `
                 -CriticalAttributesOnly:$CriticalAttributesOnly `
                 -SilenceIfNotInWarningOrCriticalState:$SilenceIfNotInWarningOrCriticalState
+            }
+        }
+        finally
+        {
+            if ($cimSessions)
+            {
+                Remove-CimSession -CimSession $cimSessions
+            }
         }
     }
+
+    # CimSession
+    elseif ($CimSession)
+    {
+        foreach ($cim in $CimSession)
+        {
+            if ($cim.TestConnection())
+            {
+                inGetDiskSmartInfo `
+                -Session $cim `
+                -ShowConvertedData:$ShowConvertedData `
+                -CriticalAttributesOnly:$CriticalAttributesOnly `
+                -SilenceIfNotInWarningOrCriticalState:$SilenceIfNotInWarningOrCriticalState
+            }
+            else
+            {
+                $Script:ErrorAccessingCimSession += $cim.ComputerName
+            }
+        }
+    }
+
+    # Localhost
     else
     {
         inGetDiskSmartInfo `
             -ShowConvertedData:$ShowConvertedData `
             -CriticalAttributesOnly:$CriticalAttributesOnly `
             -SilenceIfNotInWarningOrCriticalState:$SilenceIfNotInWarningOrCriticalState
+    }
+
+    # Error reporting
+    foreach ($e in $Script:ErrorCreatingCimSession)
+    {
+        Write-Error -Message "ComputerName: ""$($e.OriginInfo.PSComputerName)"". $($e.Exception.Message)"
+    }
+    foreach ($e in $Script:ErrorAccessingCimSession)
+    {
+        Write-Error -Message "ComputerName: ""$e"". The WinRM client cannot process the request because the CimSession cannot be accessed."
+    }
+    foreach ($e in $Script:ErrorAccessingClass)
+    {
+        Write-Error -Message "ComputerName: ""$($e.ComputerName)"", Protocol: $($e.Protocol). $($e.ErrorObject.Exception.Message)"
     }
 }
 
@@ -75,9 +115,24 @@ function inGetDiskSmartInfo
         $parameters.Add('CimSession', $Session)
     }
 
-    $disksInfo = Get-CimInstance -Namespace $namespaceWMI -ClassName $classSMARTData @parameters
-    $disksThresholds = Get-CimInstance -Namespace $namespaceWMI -ClassName $classThresholds @parameters
-    $diskDrives = Get-CimInstance -ClassName $classDiskDrive @parameters
+    try
+    {
+        $disksInfo = Get-CimInstance -Namespace $namespaceWMI -ClassName $classSMARTData @parameters  -ErrorAction Stop
+        $disksThresholds = Get-CimInstance -Namespace $namespaceWMI -ClassName $classThresholds @parameters
+        $diskDrives = Get-CimInstance -ClassName $classDiskDrive @parameters
+    }
+    catch
+    {
+        if ($DebugPreference -eq 'Continue')
+        {
+            $_
+        }
+        else
+        {
+            $Script:ErrorAccessingClass += @{ComputerName = $Session.ComputerName; Protocol = $Session.Protocol; ErrorObject = $_}
+        }
+        continue
+    }
 
     foreach ($diskInfo in $disksInfo)
     {
